dialog box renders
