+++ src/index.html
@@ -48,7 +48,8 @@
 
             <article class="friends--container">
                 <div class="friends--list"></div>
-                <div class="friends--addFriend"></div>                
+                <div class="friends--addFriend"></div> 
+                <dialog class="friend__confirmDialogBox"></dialog>             
             </article>
             
             <article class="publicChat--container"></article>
