+++ src/scripts/publicChat/publicChatHTML.js
@@ -29,7 +29,7 @@ eventHub.addEventListener("click", clickEvent => {
 export const messagesHTMLConverter = (messageObj, user) => {
     return `
     <section class="message--board">
-    <div class="message__User">Message from: <a href="#" id="addFriendFromChat--friendUserId--${messageObj.userId}">${user.username}</a></div> 
+    <div class="message__User">Message from: <a href="#" id="addFriendFromChat--friendUserId--${user.id}">${user.username}</a></div> 
     <div class="message__User" style="display: none" >${messageObj.userId}</div> 
     <div class="message__Text">${messageObj.text}</div> 
     
@@ -39,17 +39,17 @@ export const messagesHTMLConverter = (messageObj, user) => {
     `
 }
 
-// dispatch "addFriendFromChatClicked"
+// dispatch "chatAuthorClicked"
 eventHub.addEventListener("click", clickEvent => {    
     if(clickEvent.target.id.startsWith("addFriendFromChat--")) {
-        const [prompt1, prompt2, chatAuthorId] = clickEvent.target.id.split("--")
-        const customEvent = new CustomEvent("addFriendFromChatClicked", {
+        const [prompt1, prompt2, userId] = clickEvent.target.id.split("--")
+        const customEvent = new CustomEvent("chatAuthorClicked", {
             detail : {
-                friendUserId : parseInt(chatAuthorId)
+                chatAuthorId : userId
             }
         })
         eventHub.dispatchEvent(customEvent)
-        // console.log("dispatched addFriendFromChatClicked")        
+        console.log("dispatched `chatAuthorClicked`; chatAuthorId >>", customEvent.detail.chatAuthorId)        
     }
 })
